user_service/requirements.txt
Django==3.2.5
psycopg2-binary==2.9.3
djangorestframework==3.13.1
django-cors-headers==3.11.0
djangorestframework-simplejwt==5.1.0
django-environ==0.9.0